+++ lib/engineyard-serverside/lockfile_parser.rb
@@ -2,7 +2,7 @@ require 'yaml'
 module EY
   module Serverside
     class LockfileParser
-      DEFAULT = "1.0.10"
+      DEFAULT = "1.0.21"
 
       def self.default_version
         DEFAULT